getSlipStreamPools
