+++ src/viem/pool/index.ts
@@ -1,3 +1,4 @@
 export * from './pool';
 export * from './checkAutomationSupportForPool';
 export * from './getSlipStreamPools';
+export * from './getPool';
